flowerdelivery_master/requirements.txt
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
h11==0.14.0
httpcore==1.0.7
httpx==1.0.0b0
idna==3.10
pillow==11.0.0
python-telegram-bot==21.9
requests==2.32.3
rfc3986==2.0.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
